Latest commits
* note that Jinja templates are not used for now due to Google's custom license and design questions of json+Jinja
* fixes to `top_n_sigma` sampler

+++ base_sampling2/llama-addon.cpp
@@ -1520,7 +1520,7 @@ struct llama_sampler * llama_sampler_init_penalties_addon(
 // top-n-sigma
 
 struct llama_sampler_top_n_sigma {
-    const int32_t n;
+    const float n;
 };
 
 static const char * llama_sampler_top_n_sigma_name(const struct llama_sampler * /*smpl*/) {
@@ -1579,7 +1579,7 @@ static struct llama_sampler_i llama_sampler_top_n_sigma_i = {
     /* .free   = */ llama_sampler_top_n_sigma_free,
 };
 
-struct llama_sampler * llama_sampler_init_top_n_sigma(int32_t n) {
+struct llama_sampler * llama_sampler_init_top_n_sigma(float n) {
     return new llama_sampler {
         /* .iface = */ &llama_sampler_top_n_sigma_i,
         /* .ctx   = */ new llama_sampler_top_n_sigma {
